// src/lib/pose-types.ts
export type FiveElement = 'wood' | 'fire' | 'earth' | 'metal' | 'water'
// ...
export type MuscleGroup =
  | 'psoas' | 'iliacus' | 'hip-flexors' | 'hamstrings' | 'quadriceps' | 'glutes'
  | 'piriformis' | 'hip-adductors' | 'hip-abductors' | 'IT-band' | 'lumbar-spine'
  | 'thoracic-spine' | 'cervical-spine' | 'spinal-erectors' | 'chest-pectorals'
  | 'anterior-shoulder' | 'posterior-shoulder' | 'lats' | 'rhomboids' | 'trapezius'
  | 'intercostals' | 'diaphragm' | 'calves' | 'tibialis-anterior' | 'plantar-fascia'
  | 'achilles' | 'ankle-ligaments' | 'wrist-extensors' | 'obliques' | 'sacroiliac-ligaments'

export interface BreathingCues {
  entering: string
  holding: string
  exiting: string
}

export type JointAction =
  | 'flexion' | 'extension' | 'internal_rotation' | 'external_rotation'
  | 'abduction' | 'adduction' | 'compression' | 'traction'
  | 'lateral_flexion' | 'circumduction' | 'inversion' | 'eversion'

export type JointName =
  | 'lumbar' | 'thoracic' | 'cervical' | 'sacroiliac'
  | 'hip' | 'knee' | 'ankle' | 'toes'
  | 'shoulder' | 'elbow' | 'wrist'
  | 'feet' | 'neck'
// ...
export type ChakraName =
  | 'root' | 'sacral' | 'solar-plexus' | 'heart'
  | 'throat' | 'third-eye' | 'crown'